+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Updates between September 9, 2019 and [].
 1. `._build_polygons` is now multithreaded.
 2. `.create_index` is now multithreaded. Closes #65
 3. `from_dir` now supports glob strings. Closes #66
+4. `par_apply`  accepts optional keyword arguments using the `args` 
+parameter. These are passed to the applying function.
 
 ### GISExport
 1. Removed some deprecated functions that were only used for tree segmentation